+++ dimensions/2/src/io.c
@@ -2,30 +2,30 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-double** read(char *src, int *L){
-  double **data;
+long double** read(char *src, int *L){
+  long double **data;
   FILE *in;
   int it;
   in = fopen(src, "r");
   fscanf(in, "%d", L);
-  data = (double**) calloc(*L, sizeof(double*));
-  *data = (double*) calloc(*L**L, 2*sizeof(double));
+  data = (long double**) calloc(*L, sizeof(long double*));
+  *data = (long double*) calloc(*L**L, 2*sizeof(long double));
   for(it=0; it < *L**L; it++){
-    fscanf(in, "%lf", *data+2*it);
+    fscanf(in, "%Lf", *data+2*it);
     data[0][2*it+1]=0;
   }
   fclose(in);
   for(it=1; it < *L; it++)
-    data[it]=data[it-1]+*L*2*sizeof(double);
+    data[it]=data[it-1]+*L*2*sizeof(long double);
   return data;
 }
 
-void write(char *dst, int L, double **dft){
+void write(char *dst, int L, long double **dft){
   FILE* out;
   int it;
   out = fopen(dst, "w");
   fprintf(out, "%d\n", L);
   for(it=0; it < 2*L*L; it+=2)
-    fprintf(out, "(%6.2lf, %6.2lf)%c", it[*dft], (it+1)[*dft], it/2%L==L-1 ? '\n' : ' ');
+    fprintf(out, "(%6.2Lf, %6.2Lf)%c", it[*dft], (it+1)[*dft], it/2%L==L-1 ? '\n' : ' ');
   fclose(out);
 }
